fix(tradeHistory): open the peer profile when the history card's avatar is tapped
