+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/tabs/my_trades/closed/ClosedTradeListPresenter.kt
@@ -121,6 +121,9 @@ class ClosedTradeListPresenter(
             is ClosedTradeListUiAction.OnSelectTrade ->
                 _uiState.update { it.copy(selectedTradeForDetails = action.item) }
 
+            is ClosedTradeListUiAction.OnPeerProfileClick ->
+                navigateTo(NavRoute.PeerProfile(action.profileId))
+
             ClosedTradeListUiAction.OnDismissDetails ->
                 _uiState.update { it.copy(selectedTradeForDetails = null) }
